requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
numpy>=1.24.0
gunicorn>=21.0.0
